Allow filtering by site on Special:ItemsWithoutSitelinks
This patch adds a form on the top of Special:ItemsWithoutSitelinks
to allow searching only for items which have no sitelink for a
given site.

Furthermore, a test is created for the getItemsWithoutSiteLinks
method to verify that the result is ordered by the number of
sitelinks.

Things to be considered:
* Show a message when an invalid site id is entered?
* How to make clear that when no site id is entered items
  without any site link will be shown?
* <s>Order the results by the count of sitelinks the items have
  (only relevant when a site is entered obviously).</s>

Change-Id: Id4637eb701de32b2ee9ffb75847e8376258bda19