Update patch set 9
Patch Set 9:

As mentioned in irc the edit button should be moved to another place anyways (next to the "In other languages" toggle afaik) so that shouldn't be an issue.

Concerning the space which got created between label and description, I'm not sure if that is something we should touch. It's the standard space between the heading and the page's actual content defined by the Vector theme. Having the description in line with the Wikipedia heading on the right is ok imo but I have no strong opinion on that topic.

Patch-set: 9
Label: Verified=0