Update patch set 1
Patch Set 1: Code-Review+1

I tried and it works as expected. But I wonder: whats the benefit of this? Now I can search for all items that do *not* have a sitelink to dewiki, for example. What's the use of this? I think it would make more sense to have a filter per group, e.g. list all items that do not have a sitelink to Wiktionary.

Patch-set: 1
Label: Code-Review=+1