Update patch set 9
Patch Set 9:

there is siteSub and contentSub. suppose those are why the "padding" is there between the h1 and mw-content-text.

just an idea, maybe those could be used for something, or maybe better not to touch and change how description + aliases are.

imho, this all needs input from Lydia (and probably Henning)

Patch-set: 9
Label: Verified=0