Update patch set 2
Patch Set 2:

EditFilter is a recurring problem, when comes to integration with other extensions and this code sometimes not being run by jenkins / not adequately tested.

maybe this hook runner can help with the situation and allow more tests to be added for this?

Patch-set: 2
Reviewer: Aude <33@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0