Update patch set 1
Patch Set 1:

It basically requires the result to be ordered by the number of other site links the item has as stated in the commit message. However, I didn't find a nice and efficient solution for this. However, as the query is inefficient anyways we should maybe rethink the whole SQL part of this feature.

Patch-set: 1