Update patch set 1
Patch Set 1:

Doing a query for items that don't have a sitelink to site x, but have on at all, or one to site y, or one to a site of a certain group should be easy. Implementing a top n query on top of that (sorting by number of sitelinks) is probably going to be a problem.

Patch-set: 1