Update patch set 1
Patch Set 1: Code-Review-2

(1 comment)

The unused code is for measuring the cache hit rate. That's quite useful for debugging, even if currently unused.

I disagree about removing the <Change>  type parameter from type hints. That is useful information and part of the method's contract, which needs to be documented.

Patch-set: 1
Label: Code-Review=-2