Fix 32 bit integer overflow in ChangePrunerTest
Look at the intval() I removed. This casted a MediaWiki timestamp
string to integer instead of applying wfTimestamp( TS_UNIX, ...).
This is an actual bug and should have failed on all machines but only
failed in my local 32 bit PHP.

In this patch I'm changing most methods to not get and return integer
timestamps, but MediaWiki timestamp strings instead.

Bug: T96566
Change-Id: I7a7126e8ca4e369cf7a7058f634982fcb605ac9e