Adjust BotEditTest for core api change
The recentchanges output now is more explicit regarding
bot and new edits.

If bot and new props were requested in the api request, then
these params are always included in the result, instead of
only if true (with value of '').

https://gerrit.wikimedia.org/r/#/c/191103/11/includes/api/ApiQueryRecentChanges.php

also did some refactoring of the assertions
to make the code a bit more clear.

Change-Id: If0c0fe2ee70ca22da9365fd9417bac7221201c21