Update patch set 6
Patch Set 6: Code-Review-1

I'm not voting against the order thing, but I have to agree that the change to the UI/UX of the special page is barely useful for an untrained user. What the form exactly does is not clear. It looks like a filter but when entering a "filter" the list gets *longer*. It's not obvious why. I think you are trying to squish an unrelated feature into an existing special page. This probably should be an other special page or something. Maybe an API module.

Patch-set: 6
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1