Update patch set 6
Patch Set 6:

I agree that this needs more documetation. Perhaps an introduction paragraph on top of the special page. However, I disagree that this should go into a different special page. Only searching for a specific site missing does not change the intention of this special page but adds some useful functionality.

Patch-set: 6