Update patch set 6
Patch Set 6:

Users don't read documentation. After reading Hoo's comments and reconsideration I think that the new feature you are creating here should not go into a special page named "ItemsWithoutSitelinks". It barely useful anyway because of the way the special page works. Ordering is not possible. The number of sitelinks is not shown. What the user gets is basically some *random* items that happen to not have a sitelink to a certain wiki. To make this useful you need to filter the missing, untranslated pages by a lot more criteria (possibly by category). This sound like something an external tool should do.

Patch-set: 6