Update patch set 6
Patch Set 6:

When I say ordering I mean: the user can choose how to order, e.g. by item id, by label (in the users language) and so on. This is, as I said, something an external tool should do.

Displaying a number? On a special page that lists items with zero site links? If that's not an indicator that this feature does not belong here I don't know.

Patch-set: 6