Update patch set 1
Patch Set 1:

The it should be filtered before, and trigger an exception here. The RdfWriter should not just gloss over broken data. If we know this is is a problem with the Wikibase data, it should be fixed in Wikibase.

Keep in mind that purtle is intended to be a separate (for now, logical) component with no dependency on or knowledge of Wikibase.

Patch-set: 1