Update patch set 1
Patch Set 1:

What "default"? There is no "default" in this patch. This is just mapping null to an empty array before passing it to methods that don't allow null and would crash hard if called with null. Why should it be forbidden to set an optional setting to null? I don't get that part.

Patch-set: 1