Update patch set 1
Patch Set 1:

Because null is not a legal value for this setting. Just like it's not a valid value for most of the settings we have. "Sometimes" having special code to work around sloppy setup is a bad things, it hides risky or broken situations. 

If you have no badgets, set the setting to an empty array (or not at all, since the empty array is the default).

Patch-set: 1