Update patch set 1
Patch Set 1:

why should the fact that a default exists (i.e. the setting is optional) imply that null should be valid? 

You could make "null" mean "use the default", but that would be a breaking change to our config system. You would need special handling for null values (look up the default). And you would no longer be able to override a non-null default with null.

And you still not apply defaults "inline" when the setting is used. 

Imagine we'd want to change the default for some reason (and let's assume we follow Kant and want to apply the same rules and reasoning for *all* settings, not just for this special case). Then we'd have to hunt down all places where the setting is used, and see what kind of special handling these places have for null (or false. or an empty string. or whatever else). 

We currently have a simple rule, which I suggest we should just follow for all settings:
* define defaults in the defaults file
* override settings in the LocalSettings.php file
* when using settings, assume they are always present and well defined

Patch-set: 1