Cache supported entity data formats
As generating that list is very expensive.

The is needed for T96298 as we want to use the Entity data
formats on every page view there.

As a side effect this makes Special:EntityData a little faster
(I get about 13% for content negotiation only and about 5% when
doing content negotiation and then loading a mid sized entity as
ttl).

Change-Id: If7d137a61215f31d2a110b1272b99e31984d7dbd