Update patch set 6
Patch Set 6:

Sorry, but this is not a valid argument because [[Special:EntitiesWithoutLabel]] can not be used to list entities that do not have *any* label. It's *always* filtered by language. My argument still applies: I don't think it's a good idea to squish two different features into a single special page.

Patch-set: 6