Remove unnecessary intval() and add missing (int) casts
intval() is about 5 times slower than a type cast and should be avoided
if not necessary. Comparisons with == do not need an explicit type cast,
this happens automatically.

This patch also adds some (int) type casts in places where it can turn
into a security issue if a string is used as part of a WHERE clause.

Change-Id: I21bd5a8b833e908c36d09b10d29141c91594ef30