Update patch set 2
Patch Set 2:

(4 comments)

The code looks ok on a first glance, but I didn't check closely. There is one naming issue that bothers me.

In the bigger picture, I'm wondering whether we really want to cache on this level. I think info about alternative formats should go into the ParserOutput, so we don't have to re-generate it on every page view. On the other hand, the formats are the same for all entity pages...

Generally, I'm worried that we'll hit whatever cache will be used for caching the available formats will be hit on every page view.

Patch-set: 2