Update patch set 2
Patch Set 2: Code-Review-1

To get an idea about the performance impact on HHVM: https://gist.github.com/mariushoch/2067b450a79653cdb24f (I don't know for sure which of these numbers are more likely what we're going to see during a web request... potentially rather the lower ones than the bigger ones as a huge fraction of the time spent goes into loading ApiMain which is only needed in CLI mode). On Zend the impact is way larger (even on 5.6).

Because of the above I'm inclined to abandon this (maybe the factored out EntityDataFormatProvider would be worth keeping because it's nicer to have that separate, I don't know).

Patch-set: 2
Label: Code-Review=-1