[DNM] Use ByPropertyIdMap in ChangeOpClaim
This changes the way how ChangeOpClaim and wbsetclaim
in the Api deal with indices. Instead of having a single
index per statement, it now introduces to indices, one for
a statement within a property group and one for the property
group itself.

Currently, indexing is not supported in the ui so that does
not require any changes. (Stats for usage of the index parameter
still needed to check if b/c has to be added in some ways.)

Bug: T98375
Change-Id: I74c2e48f0cde398e6d6eb9b0aaf50870d8284c22