[DNM] Use ByPropertyIdMap in ChangeOpClaim
This changes the way how ChangeOpClaim and wbsetclaim
in the Api deal with indices. Instead of having a single
index per statement, it now introduces two indices, one for
a statement within a property group and one for the property
group itself.

Currently, indexing is not supported in the ui so that does
not require any changes. No usage has been found in the last
24 hours so this parameter can be considered unused and we
don't have to care about b/c and broken tools/bots.

Bug: T98375
Change-Id: I74c2e48f0cde398e6d6eb9b0aaf50870d8284c22