Track label usage including fallback chain.
This moves the tracking logic into a decorator implementation of
SnakFormatter. This moves the tracking closer to the actual usage in
the ValueFormatter, and reduces the number of places that need
to care about usage tracking.

This change includes several modifications to test cases to ensure that
the SnakFormatter decorator is applied and configured correctly.

Change-Id: I0fe2b184e8c8f096af37bcff6a7f6ecd9355d663