Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

Good to see somebody working on this. The patch is fine. But I think the possibility to display the entity id in labelview is a useful feature and should not be removed, even if we do not use it in the moment. Other team members may disagree. I may be convinced. But for now:

1. There are at least 2 unrelateted bugfixes. Please move them to a separate patch.

2. Make the entityId parameter optional/nullable but do not remove it completely.

Patch-set: 2
Label: Code-Review=-1