Update patch set 2
Patch Set 2:

(1 comment)

I don't see how this feature is dangerous and must be destroyed. The fact that you can identify an entity by it's id is not something that "just [...] there because we can" but because it was actually used in the UI just recently and may be of use in the not so distant future again or in alternative UIs that want to reuse some of our widgets.

Patch-set: 2