Update patch set 2
Patch Set 2:

As said, it is really easy to add the feature again and show the entity id in the labelview. However, I'm really against having settings which were used in code which got removed still being alive. If we currently have no use for them, there is no reason to keep them. Furthermore, I see no point why this "may be of use in the not so distant future again" according to the current mockups provided.

Patch-set: 2