Improve detection of non-content parser runs.
On ParserAfterParse, we pull in sitelinks from the repo and inject
them into the ParserOutputObject. We should only do this when parsing
actual page content. Detecting when this is the case is not trivial
from inside the ParserAfterParse handler function.

In the past, we used $parser->getInterfaceMessage() === false. But that
would pass for any parse that does not use the user interface message.
A number of interface components or actions don't, for instance, the edit
toolbar on the edit page.

This change fixes that by now checking $parser->getRevisionId() !== null;
The revision ID is set whenever actual page content is parsed, and only
then, apparently. Which is exactly what we want.

Note that this means that we no longer inject sitelinks during edit
previews.

Change-Id: I30d12214b04626a903e125d1434b3d37303205a7