Make config script more robust.
Make it independent of the current state of LocalSettings.php: Either add
opening and closing PHP tags when the file is closed with one, or don't.

This is necessary because I plan to change when this script is executed by our
CI. At that point the CIs LocalSettings.php is not closed with a PHP tag, where
as when it is currently execute there is a closing PHP tag.

Bug: T97529
Change-Id: Ie5611fd24c0f245b2aceb48610a866c61cf8a4e1
(cherry picked from commit daeed233b32d4f4da6a2eeb51d1beb6cf8d8902c)