Update patch set 7
Patch Set 7: Code-Review-1

(6 comments)

I'm also concerned about the changed meaning of the "index" parameter. I'm really not sure if this is a good idea, even if I understand the idea and agree that it does have benefits.

The problem is: The StatementList in the Entity is not guaranteed to be grouped by anything. A single "index" parameter does not add such assumptions and does not requires the user to know about grouping. With the changed indexing in this patch users must be aware of the goruping. I'm not sure if this is helpful.

Can you please split this into two patches? The first reworks the code without changing the meaning of the "index" parameter, the second changes the index parameters.

-1 because of missing/not updated parameter documentation.

Patch-set: 7
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1