Add missing newlines between "namespace" and "use" blocks
This fixes a single style guide issue accross the whole Wikibase.git
code base: there must be a blank line between "namespace" and "use"
(according to phpcs) and there should be a blank line before "namespace"
(I could not find a phpcs rule for that).

This is split from If24d3d1 to make it easier to review.

Change-Id: Icf93d0219f73a5901b7e9eebd4bc3ce60a7d3d64