Use LinksUpdateComplete hook for usage tracking updates.
In order to update usage tracking upon template edits,
this change uses the LinksUpdateComplete instead of the ArticleEditUpdates
hook we used before.

ArticleEditUpdates is only triggered when the page itself is edited,
but we need to also update usage tracking when templates transcluded
on the page change. LinksUpdateComplete is triggered whenever MediaWiki
decides it needs to re-evauate which links are present on a (rendered)
page, which should be exactly when we want to also update usage tracking
(entity usages are essentially links, just like template or image usages).

Bug: T99510
Change-Id: Iee2f0ca7b7adb1482c0fc6f4a0230da75d5beefd