Update patch set 12
Patch Set 12:

updated the docs

also decided that defaulting to the connected item is somewhat silly for mw.wikibase.sitelink since it would just return the page title of the current page. (for which there certainly are better, cheaper, other ways to do)

for label and description, having the default makes sense.

Patch-set: 12
Label: Verified=0