Update patch set 1
Patch Set 1:

@thiemo: ParserAfterParse is called whenever any wikitext is parsed for any reason, including interface text, system messages, old revisions, diff views, etc. We tried some hacks to only trigger the expensive operations on "real" parsing of page content, but that was never reliable.

Using the ContentAlterParserOutput hook, our code gets called only when page content is being rendered, and it's called also for non-wikitext content. It gets executed for previews, but not for system messages parses. Which is what we want.

The "alter" means that the hook allows us to modify the "normal" parser output, in contrast to the ContentParserOutput hook which would allow us to supply a completely different ParserOutput object, skipping the normal process of parsing.

Patch-set: 1
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0