Allow filtering by property in wikibase.renderSnaks
This patch adds an optional paramter to the wikibase.renderSnaks
lua method to allow filtering the passed snaks by a given property
label or id. This is needed when only parts of a snak list (eg. in
a reference or in qualifiers) should be rendered, or if some snaks
for a given property should be rendered differently.

Bug: T100899
Change-Id: I56e843d226f25b54aef4673fe936adc849cc9939