Undeprecate term setters in Item and Property
Right now code as follows is deprecated:

$item = new Item();
$item->setLabel( 'en', 'foo' );
$item->setDescription( 'de', 'bar' );

This is caused by these setters being deprecated in Entity, which
was done to deprecate code assuming all entities have these terms,
and to discourage bad interface seggergation. The above code suffers
from neither these problems, and is a style often used in tests,
and has no clean alternative due to PHPs lack of named parameters.
Hence this undeprecation in Item and Property, keeping the one in
Entity.