Update patch set 11
Patch Set 11:

once we write these new usages to the database (e.g. 'L.ku'), then in the event of some problem with deployment, the deployment branch has to be reverted to the previous one, then the existing code causes stuff like saving a page to fail.

Exception encountered, of type "InvalidArgumentException"
[cd52f46b] /w/index.php?title=Bangkok&action=submit InvalidArgumentException from line 96 of /var/www/wiki/w/extensions/Wikidata/extensions/Wikibase/client/includes/Usage/EntityUsage.php: $aspect must use one of the XXX_USAGE constants!

not so much necessarily a problem of this patch, but would like the previously deployed code to somehow be more robust for such a situation. (as this situation did occur last week and does happen)

Patch-set: 11