Update patch set 13
Patch Set 13:

i am not sure why or if it is related, but I seem to get a lot of "lock wait timeout" errors with this patch when editing a page in the client or sometimes even loading a client page. (it could be related to https://phabricator.wikimedia.org/T100577)

with master of wikibase, I am not encountering the issue, though should poke more to be sure.

Patch-set: 13