Update patch set 11
Patch Set 11: Code-Review-1

last comment for now...

we also need to bump the PARSER_VERSION constant in EntityHandler, since this is probably a breaking change for parser output that necessitates purging caches.

and I would like tobi or someone to run browser tests with this patch.  it might be that the tests need to be adjusted for this change, which someone can do in a separate task. (but important to know)

Patch-set: 11
Reviewer: Aude <33@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1