Fix behaviour of EntityRevisionLookup::getLatestRevisionId implementations
The MockRepository one threw an UnresolvedRedirectException
if the entity it was called with is a redirect, but the "real"
implementation in WikiPageEntityRevisionLookup doesn't do that
(and can't really, without having to query more data per default).

In order to fix this, I changed all implementations to work with
redirects in a consistent manner (like it was documented in the
interface) and adopted all callers that relied on the implemetation
in MockRepository.

Bug: T101625
Change-Id: I8d05f1fb093216cf30168ae52a27ac8dd53e3d04