Update patch set 3
Patch Set 3: Code-Review-1

I have to say that it makes me rather uncomfortable that EntityRevisionLookup::getLatestRevisionId shoud return the revision ID of a redirect. A redirect isn't an Entity, and EntityRevision cannot represent a redirect. This means the return value of getLatestRevisionId may not be usable with getEntityRevision. 

On the other hand, throwing an UnresolvedRedirectException isn't technically needed, and Hoo is right that it would require us to load and deserialize the revision in question to find the target ID to include in the Exception.

Perhaps it would be best to simply return false here. The interface spec sais "or false if there is no such entity". A redirect is not an entity. So, according to the spect, thsi should return false for redirects. 

Are there any practical issues with that approach?

Patch-set: 3
Label: Code-Review=-1