Update patch set 44
Patch Set 44:

Jeroen De Dauw, if you think we should improve the extension registration mechanism offered by core or CI related to it, please file a task or use another place to discuss this that will not be ignored.

This is a commit that does not change how core expects extensions to register, so as far as I understand your comment is not requesting changes to this commit.

Patch-set: 44
Label: Verified=0