Rework EntityTermsViewTest to not break because of core change
This patch fixes an issue that currently breaks all tests on Travis:
https://travis-ci.org/wikimedia/mediawiki-extensions-Wikibase/
For example:
https://travis-ci.org/wikimedia/mediawiki-extensions-Wikibase/jobs/66086125

This change removes an aspect that was tested before: The last test checked
if HTML injected via a message (from the MediaWiki: namespace) is escaped
or not. This is not tested any more. This is done on purpose. It's not
really relevant. Users that can inject HTML this way can also inject it in
a gazilion other places. It's not the responsibility of this specific test
to make sure HTML injections aren't possible.

Note that this is the only (!) place in the whole Wikibase code base
(including all components) that made use of this MessageCache trick. There
is probably an other way in core to do the same but I did not found out.

Change-Id: I80df6858e3ab3b58f721881ded82a68af4658c70