Update patch set 6
Patch Set 6:

I also thought about this. Do you have a better suggestion for the name?

The reason why we want to introduce this class is that lots of code does the same for fetching terms of an entity and we want to provide a convenient way to do so.

Patch-set: 6