Update patch set 3
Patch Set 3:

But we don't really need the revision info. We just need to know whether the revision is a redirect, and we only need that for the current revision. So we just need page_is_redirect, which may well already be loaded in the LinkCache (which, of course, is quite horrible).

Patch-set: 3