Make "site link" two words in all messages
This was raised as a question in I64d91e2. Is "sitelink" one word or
should it always be written as "site link"? I suggest the later. I
would love to get a +/-1 from Lydia, Aude and Addshore and decide
based on their votes.

Note that currently there are much, much more occurences of "sitelink"
as one word. This patch will change that.

Change-Id: I4ed29e82525b3c3357ddbd5035e05b649f8c4ec9