Update patch set 2
Patch Set 2:

> Bene, please create a Phabricator ticket if you think it's useful.

Is this the way it works? Really??

Furthermore, I personally like the term "sitelink" more because it became an established term on Wikidata. However, I think this is a very subjective and we should not invest much time of discussion in this. Also, this is currently not on any sprint.

Patch-set: 2