Update patch set 3
Patch Set 3: Code-Review+1

The code looks good on a first glance, but didn't manually test it, and the tests need fixing.

Note: this patch keep s the old sort order (namespace+title). That should be efficient enough, and work for paging. If we want to sort py page id instead, that should be a separate patch.

Patch-set: 3
Label: Code-Review=+1