Update patch set 2
Patch Set 2:

I'm neutral. Both are OK. "Sitelink" in one word seems to be much more common both in the code in the wikidata.org site, but it's not impossible to make it consistent if that's what the community wants.

Patch-set: 2
Reviewer: Amire80 <54@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0