Update patch set 10
Patch Set 10:

@thiemo: what's your -1 on PS9 actually for? Did you see my replies? 

I agree that the hardcoded URI for Q199 is nasty. Can you think of a better way to do this? Basically, we have to know all the special unit ids that mean "no unit" somehow. Ideally before parsing (at least for "" and "1", since these are not valid URIs and would fail to parse).

Patch-set: 10