Update patch set 2
Patch Set 2: Code-Review-1

Ugh, offset. Can't we just drop paging? This is a search interface (based on fuzzy matches and ranking), not a query interface. Paging on a ranked result is not terribly useful (who goes to page two of google results?) and always expensive (ranking is dynamic). 

At the very least, there needs to be a hard limit to the offset. 1000 perhaps. CR-1 for that, since it's a potential DoS vector.

Patch-set: 2
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1