Update patch set 12
Patch Set 12: Code-Review-1

Restoring my -1.

The design of the EntityLabelUnitFormatter and it's tests. A confusing mixture of stuff that should be parseable and stuff that should not but is. From the tests it's unclear what this thing is supposed to do and what it's purpose is. Parse entity URIs or entity ids? It looks like it parses both. This can't be right.

The fact that the formatter accepts any EntityIdParser, while in reality it won't work with any implementation but one.

Patch-set: 12
Label: Code-Review=-1