Update patch set 2
Patch Set 2:

This is only refactoring and doesn't change behavior. The renames make it more obvious that units should be items, not other entity types, but the code still accepts (and still uses) parsers that accept all entity types.

See https://github.com/wmde/WikibaseDataModel/pull/503 for that parser, we currently do not have it. We could use an ugly new DispatchingEntityIdParser( array( ... ) ) thing but I really do not want to abuse it for this purpose.

Patch-set: 2
Label: Verified=0