Update patch set 9
Patch Set 9: Code-Review-1

When I edit a reference that contains a deleted property, and click cancel, the reference's snak is rendered as:

undefined (Deleted property)
[object Object]

This is not really a new issue. Before this patch it was actually worse, the reference's snak disappeared. So we can still merge this and fix the bug I described later. Please tell me what you like more.

Patch-set: 9