Rework EntityLabelUnitFormatter into ItemLabelUnitFormatter
This is an alternate proposal. One way to describe what this patch
does is: it inlines the entity id parser into the unit formatter and
does not use an actual EntityIdParser instance for that any more.
I do not think this is wrong. The "parser" part is super-trivial anyway.
Why does it need an extra class?

This does not get rid of the SuffixEntityIdParser because it's still
used for usage tracking.

See https://github.com/wmde/WikibaseDataModel/pull/503 for an
alternative proposal.

Bug: T102840
Change-Id: Ifac11f50910bbe0f329fb1ed4191aa4395fcfff5