Update patch set 1
Patch Set 1:

I don't see why this should be limited to Items, or why the formatter should know or care. Using Properties doesn't make sense, but why restrict the *formatting* to Items, if there may be other item types that we may want to allow?

For instance, we may decide to introduce a dedicated entity type for units. We actually considered this option for quite a while.

Basically, the validator should be responsible for deciding what kinds of URIs can be used as a unit. The formatter should just make a best effort to show a human readable form of the unit ID. I see no reason to add complexity just to prevent the formatter prom making property URIs pretty.

Patch-set: 1