Update patch set 1
Patch Set 1:

Things that are known to be wrong should not be prettyfied. I remember you using this as an argument multiple times.

We know units refering to non-item namespaces are wrong.

If there is a unit namespace, we know all other namespaces can not be used as units.

This patch actually *decreases* complexity. I don't understand what you mean when you are using this as an argument *against* this patch.

Patch-set: 1