Update patch set 1
Patch Set 1:

How does this decrease complexity?

It is true that I often argue that breakage should not be *hidden*. The question here is really which part of the system should know about the restrictions.

I'm fine with using an EntityIdParser that will not work on properties (a special instance of DispatchingEntityIdParser would work). I'm not ok with hardcoding the assumption that units will always be items.

Patch-set: 1
Label: Code-Review=0