Update patch set 2
Patch Set 2: Code-Review-2

The base URI should be used as configured. Messing with it will lead to inconsistencies with other places where the base URI is used.

I agree that the base URI should always use http://, not https://, but that should be enforced *everywhere* we are using the concept URI, not just for RDF. 

Basically, if you don't want the base URI to use https://, don't set it to use https://!

Patch-set: 2