Fix behaviour of EntityRevisionLookup::getLatestRevisionId implementations
The MockRepository one threw an UnresolvedRedirectException
if the entity it was called with is a redirect, but the "real"
implementation in WikiPageEntityRevisionLookup doesn't do that.

In order to fix this, I changed all implementations to work with
redirects in a consistent manner by returning false for them.
All callers that relied on the implemetation in MockRepository
have been changed to use EntityRedirectLookup.

Bug: T101625
Change-Id: I8d05f1fb093216cf30168ae52a27ac8dd53e3d04