Update patch set 2
Patch Set 2:

But we don't set it, data URI is derived from canonical URI, which is all https now. Should we change it and have another configuration for data URI?

Also, canonical URI is by default described as:

		'conceptBaseUri' => function() {
			$uri = $GLOBALS['wgServer'];
			$uri = preg_replace( '!^//!', 'http://', $uri );
			$uri = $uri . '/entity/';

			return $uri;
		},

which means if somebody sets wgServer to https (which makes sense), their default concept URI would be https too. Should we change this code?

Patch-set: 2