Update patch set 1
Patch Set 1:

Oh. I think I just remembered the fundamental problem with this refactoring, the snag that kept me from doing it last time. TermIndex::getMatchingIDs does ranking (programmatically). SearchInteractor::searchForTerms relies on TermIndex::getMatchingTerms, which doesn't do ranking.

getMatchingIDs() is implemented in a very nasty way, and should really go away. What it does should probably be re-implemented in the TermSearchInteractor. I'm afraid there isn't a very nice way to implement this on top of the current terms table, so it would probably be the logic that is used by getMatchingIDs for now. When we implement this interface based on Elastic, we can do better, I hope.

Patch-set: 1