Move over Diff code from DataModel
This is a move with of the DataModel Diff code with minimal changes.
The namespace has been changed from `Wikibase\DataModel\Entity\Diff`
to `Wikibase\DataModel\Services\Diff`. Furthermore, the package
private classes have been put into a `Internal` sub namespace to
make the visibility more distinct.

This change allows for removing the Diff code, and the dependency
on the Diff library, from Wikibase DataModel in its next major
release.

The objects representing the diffs themselves contain some technical
debt. With this move we can solve this when we see fit without
having to create a breaking release for Wikibase DataModel itself.