Update patch set 1
Patch Set 1:

@Daniel As far as I can tell in the previous version all 'ranking' is done in getRankedMatches, because of this the ranking is simply:
 1 - Direct ID match
 2 - Matches in exact user search language
 3 - Matches in fallback languages (in some order)
This class now still retrieves the direct ID match first and appends what is received from the TermSearchInteractor. This inter actor first looks for matches in the search language and then in fallback languages (same order as previously done in this class).

We could add an additional level of ranking based on the number of term matches per entity (which is info the interactor and this class have)

@Jeroen will do

Patch-set: 1