Sort output of Special:UnconnectedPages by page ID (descending)
After the refactoring that based SpecialUnconnectedPages on QueryPage,
the ORDER BY clause was already ignored and overwritten by getOrderFields(),
causing the result to be sorted by page ID. However, since sortDescending()
was overwritten to return false, the result was shown with ascending page ids,
showing always the same (oldest) pages first. For Special:UnconnectedPages,
it is more useful to see new unconnected pages first.

This patch removes the redundant sortDescending() and getOrderFields() methods,
as well as the obsolete ORDER BY option.

Bug: T102817
Change-Id: I89692d9c0f0b76acd9b283e7276e8670929cc14d