Update patch set 6
Patch Set 6: Code-Review-1

(3 comments)

Nothing actually wrong with the code. It does what it is supposed to do. But I suggest to remove the not needed setStatements call, the not needed return value and to inline the few remaining lines from the private function. -1 to get your attention. Feel free to disagree. We can also merge this and refactor it later.

Patch-set: 6
Label: Code-Review=-1