Add "composer test" command for Jenkins
Originally I copied this line from core. It used a parameter and you
had to provide a directory when calling "composer phpcs .". Now the
line is mostly copied from the Vector skin. This allows to run
"composer test" on Jenkins with no extra parameter.

The order of the phpcs parameters follows the order given via "phpcs -?".

Change-Id: I32e2e231521ac2f78aa9d08bb42c3dd37911e537