Update patch set 1
Patch Set 1: Code-Review+1

It's supposed to call whatever phpcs is finds. I don't have it in vendor in my local environment. Why should I? How is this a reason to not give developers the possibility to type "composer test" in addition to "composer phpcs"? Please note that the detail you are using to block this patch is not introduced by this patch. So I do not see a valid reason for the -1 and not way forward. What now?

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1