Update patch set 1
Patch Set 1:

When a new patch set gets uploaded, I usually compare additionally to the "normal" comparison to master the comparison to the last patch set (select Reference Version: PS 1). There I saw that lots of files changed (i18n stuff etc.) so PS 2 clearly was a rebase. I might've checked the files touched in the patch to see if they changed as well but I assumed that rebases and changed functionality should be separated.

(I learned that lesson in my first days of gerrit development, see https://gerrit.wikimedia.org/r/#/c/89859/ the comment by Siebrand: "@Bene: It looks like you're rebasing and making code changes in the same patch set. That makes it hard to compare the state the of the patch set the comments were made in with the current state. If at all possible, only rebase just before merging, or at least do not make changes in the patch set that is the rebase.")

Patch-set: 1