Update patch set 9
Patch Set 9:

Note: The 3 files not starting with Mw... are not actually related to this patch. But they are in this patch because I'm doing the exact same refactoring in all tests. However, if you think this refactoring should be in an other patch then please leave a -1 and I will split this.

Patch-set: 9
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0