Update patch set 2
Patch Set 2:

(7 comments)

I did an actual code review. Everything is fine except for the places I marked with comments.

As said before this patch is in a state that does not allow us to enable any rule but one. I prefer to enable one rule per patch. I will happily continue working on this next week. I will rebase this patch on top of mine if this is ok for you. Or start creating patches that add one rule each. The list of candidates is in the comments above. :-)

Patch-set: 2