Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

I can't tell if this is the best solution we could aim for, but the code in this patch is fine and can not break anything. Can be improved later if necessary.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1