Update patch set 1
Patch Set 1:

The use case for the "aliases" field is to show *why* a specific match is shown, in case the label doesn't match the search string. This patch removed that behavior from our UI completely, as far as I can tell.

Note that the display label in the "label" field may not be in the search language, and this needs to be reported as an alias to be shown.

Patch-set: 1