Do not trim the T in TimeDetailsFormatter
I did that in I27861d4. I think I was wrong. Originally the idea of the
timezone was similar to the calendar: the timestamp string is *always*
Gregorian and *always* UTC (Z). The calendar option was meant as a
formatter hint. Same for the timezone option.

We are changing the meaning of the calendar option now. I think we will
do the same with the timezone option.

I suggest to show the timestamp string as it is (including the Z) no
matter what. This makes this more transparent/visible.

Change-Id: Id453e9e9f57a3bfa246b83aa1dd17e42017c2313