On page views, update usage tracking via the job queue.
In some cases, we need to update the usage tracking table during normal
page views. In this case, we should not write to the database directly,
but schedule a job instead.

FIXME: I was unable to verify that this works. I see the jobs in the
database, but I cannot get them to run for some reason. runJobs.php
does nothing.

Bug: T103429
Change-Id: Ied1db3f2ab802ac7d7638b63844598374bd59812